Add runnable part
In other words,
this commit adds a runnable section to the Java Interpreter
which allows the user to access a series of tools available
through this interpreter.

At the moment this is still a Work In Progress (as you can
see the GUI is still incomplete and not well-designed,
since I used the code for Minecraft Uncovered's GUI), but
it is functional: the SKS parser can be accessed.

The same applies to icons and logos: I am too tired to
start working in Photoshop right now...

Signed-off by TheSilkMiner