Add Blacklist-Whitelist listener
A custom language used to mark whitelist or blacklist
script files.

Since a lot of people would use this type of language
(at least I think), it makes sense to have a listener
directly bounded into the implementation.

The syntax is really simple: + for whitelist, - for
blacklist, # for comments.

Signed-off by TheSilkMiner