Add services and service implementation to SKL interpreter
This allows for an easier way of parsing and/or writing
databases without having to write a custom parser: just
supply a custom service implementation and the needed
classes.

Maybe more work is needed than a custom parser rewrite,
but this is more flexible anyway.

Also provides a new parser, rewritten completely from
scratch in order to natively support services,
callbacks and custom types acceptance (is that even a
word?). The only way to use it is relying directly on
implementation, but soon enough a new highly customizable
service will be provided in the interpreter's extension
package (probably skdx or interpreterx, still deciding).

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>