Add ParserEx extension
First ever extension written for the SKD parser, it provides
a better parser and new tag types, everything written from
scratch to provide better integration and optimization.

Refer to the specific documentation for more information
(and, if it is not clear enough, tell me).

By the way, this extension exploites a little bug, but...
nothing major.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>