Use Licenser instead of Gradle License Plugin
Minecrell's version brings a few advantages over the Maven-derived
plugin we used before. The first and most important one is surely
the weight of the entire plugin: Minecrell's version does not rely
on external dependencies and so it provides a quicker download
experience, especially for CI environments such as Travis, where
build speed is important. Also, support for Kotlin files is provided
out-of-the-box, whereas the previous plugin's documentation is
foggy in this area. At least, when I looked up whether it could be
used with Kotlin files, not a lot of information came up. One small
disadvantage of this move is that we lose strict checks, but this
does not seem like a problem. Third party libraries are instead
first-class citizens and as such can be specified easily in the
plugin configuration. I don't know whether this is useful or not
for us, but it seems something worth looking into.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>